Initial design of Formatter
`HSpec` formatters are not powerful enough to represent `JUnit`. There
are a number of stats and other details that are missing, but the main
issue is that `JUnit` does not allow you to nest `<testsuite />` elements.
The base formatter outputs a normalized schema with nested `<testsuite />`
elements. A denormalization pass is then done to collapse these elements
into valid `JUnit`.