Utilize the new Test.Hspec.Core.Format module
The new format module is more flexible than the previous `writeLn`
formatting system. This is a scratch implementation of a new version of
the package that no longer needs all the indirection of printing,
parsing, and re-rendering a junit.xml file. Large swaths of code have
been deleted. What is left to do is to validate the output. This package
currently lacks tests, so improvements can be made there.