Add tests, timestamping, and valid IDs
This commit adds a golden test to the suite. This still doesn't function
as it needs to normalize timestamps and other values to work properly,
but it is a start towards coverage.

As well timestamps now report the start time of the suite and IDs follow
the incrementing ID schema prescribed for JUnit.