Add missing Formatter field in hspec-2.7.10
hspec-2.7.10 is used by LTS-18.x, so without this we are failing with
that resolver.

Our main branch has moved onto hspec-2.8, and uses an entirely different
mechanism, so we have to apply this fix as a hotfix onto v1.0.0.4 that
will never merge.

It adds the missing record field as a no-op. The hooks for examples
ending open and close their own tags, so this should be the required
behavior. It may have made more sense to let the tag open on
exampleStarted, but I'm not 100% certain, so I don't want to risk
breaking things and am electing to make the more minimal change.