Introduce JUnitConfig
Our modules were mis-named ("HSpec" instead of "Hspec"). This presents
an opportunity to change our interface in a minor version bump by doing
so in a (new) correctly-named module space.

This patch does that, and moves from positional arguments to a
JUnitConfig record. The record supports the same behavior but is more
flexible and has been written in an encapsulated style to start. Names
were chosen semi-arbitrary to match jest-junit. Documentation will
follow soon.

This configuration object will be useful as we evolve the reporter's
features; for example we will be adding the ability to specify prefixes
for reported paths (e.g. to support monorepositories). Having this
record will make that a backwards-compatible extension.

The HSpec-named modules have remained as pass-through to the new
name-space, and will be deprecated in the next commit.