Fall back to itemLocation
itemLocation is almost always present, but in the case of a positioned
result it differs from that in meaningful way:

    it "does a thing" $ do      // <-- itemLocation
      x <- this
      y <- that
      x `shouldBe` y            // <-- itemResultLocation

Therefore, we should prefer itemResultLocation but can fallback to
itemLocation if it's missing. This gets a best-effort location of the
"it" for cases without positioned results (e.g. Success), as visible in
the golden.xml diff.