Add file and line attributes
Location information isn't an official part of the JUnit XML spec, so
various conventions have emerged for reporting locations such as within
the name attribute, inferring from the classname using project structure
assumptions, or including it in the content of the testcase node (as we
do).

At least one tool* that reads JUnit reports will look for file and
line attributes if present on the testcase node. To better support this
and any similarly-behaving tools, let's add them.

*mikepenz/action-junit-report

A note about Item locations:

itemLocation is almost always present, but in the case of a positioned
result it differs from that in meaningful way:

    it "does a thing" $ do      // <-- itemLocation
      x <- this
      y <- that
      x `shouldBe` y            // <-- itemResultLocation

Therefore, we should prefer itemResultLocation but can fallback to
itemLocation if it's missing. This gets a best-effort location of the
"it" for cases without positioned results (e.g. Success), as visible in
the golden.xml diff.