Set STACK_YAML to our oldest resolver when releasing
Using --pvp-bounds=lower (which this does) sets any unspecified lower
bounds to the version present in the active resolver. Since we (now) CI
on older GHCs through older resolver sets, we should use the oldest one
we have for this operation, so we set the lowest lower bounds we can
(while still setting bounds to things we actively CI with).